Issue #549 - Fix build-warning for qsort.o
Fixed by applying the patch proposed in
https://mail.gnu.org/archive/html/avr-libc-dev/2002-08/msg00054.html

	* libc/stdlib/qsort.c: Fix build warning.
	* libc/stdlib/Files.am (qsort.o): Remove -Wno-sign-compare flag.

+++ libc/stdlib/qsort.c
@@ -137,11 +137,11 @@ loop:
 	pn = (char *)a + n * es;
 	r = min(pa - (char *)a, pb - pa);
 	vecswap(a, pb - r, r);
-	r = min(pd - pc, pn - pd - es);
+	r = min(pd - pc, (int)(pn - pd - es));
 	vecswap(pb, pn - r, r);
-	if ((r = pb - pa) > es)
+	if ((r = pb - pa) > (int)es)
 		qsort(a, r / es, es, cmp);
-	if ((r = pd - pc) > es) {
+	if ((r = pd - pc) > (int)es) {
 		/* Iterate rather than recurse to save stack space */
 		a = pn - r;
 		n = r / es;
